+++ nodes/knxUltimate-config.js
@@ -77,8 +77,8 @@ module.exports = (RED) => {
         node.KNXEthInterfaceManuallyInput = typeof config.KNXEthInterfaceManuallyInput === "undefined" ? "" : config.KNXEthInterfaceManuallyInput; // If you manually set the interface name, it will be wrote here
         node.statusDisplayLastUpdate = config.statusDisplayLastUpdate || true;
         node.statusDisplayDeviceNameWhenALL = config.statusDisplayDeviceNameWhenALL || false;
-        node.statusDisplayDataPoint = config.statusDisplayDataPoint || false;
-        node.telegramsQueue = [];  // 02/01/2020 Queue containing telegrams
+        node.statusDisplayDataPoint = typeof config.statusDisplayDataPoint === "undefined" ? false : config.statusDisplayDataPoint;
+        node.telegramsQueue = [];  // 02/01/2020 Queue containing telegrams 
         node.timerSendTelegramFromQueue = setInterval(handleTelegramQueue, 50); // 02/01/2020 Start the timer that handles the queue of telegrams
         node.timerDoInitialRead = null; // 17/02/2020 Timer (timeout) to do initial read of all nodes requesting initial read, after all nodes have been registered to the sercer
         node.stopETSImportIfNoDatapoint = typeof config.stopETSImportIfNoDatapoint === "undefined" ? "stop" : config.stopETSImportIfNoDatapoint; // 09/01/2020 Stop or Skip the import if a group address has unset datapoint
@@ -255,12 +255,12 @@ module.exports = (RED) => {
                             for (let index = 0; index < node.csv.length; index++) {
                                 const element = node.csv[index];
                                 if (readHistory.includes(element.ga)) return;
-                                node.writeQueueAdd({ grpaddr: element.ga, payload: "", dpt: "", outputtype: "read" });
+                                node.writeQueueAdd({ grpaddr: element.ga, payload: "", dpt: "", outputtype: "read", nodecallerid: element.id });
                                 readHistory.push(element.ga)
                             }
                         } else {
                             if (readHistory.includes(oClient.topic)) return;
-                            node.writeQueueAdd({ grpaddr: oClient.topic, payload: "", dpt: "", outputtype: "read" });
+                            node.writeQueueAdd({ grpaddr: oClient.topic, payload: "", dpt: "", outputtype: "read", nodecallerid: oClient.id });
                             readHistory.push(oClient.topic)
                         }
                     })
@@ -480,25 +480,58 @@ module.exports = (RED) => {
 
         // 02/01/2020 All sent messages are queued, to allow at least 50 milliseconds between each telegram sent to the bus
         node.writeQueueAdd = _oKNXMessage => {
-            // _oKNXMessage is { grpaddr, payload,dpt,outputtype (write or response)}
-            node.telegramsQueue.unshift(_oKNXMessage); // Add _oKNXMessage as first in the buffer
+            // _oKNXMessage is { grpaddr, payload,dpt,outputtype (write or response),nodecallerid (id of the node sending adding the telegram to the queue)}
+            node.telegramsQueue.unshift(_oKNXMessage); // Add _oKNXMessage as first in the queue pile
         }
 
         function handleTelegramQueue() {
             if (node.knxConnection) {
-                if (node.telegramsQueue.length == 0) {
+
+                if (typeof node.lockHandleTelegramQueue !== "undefined" && node.lockHandleTelegramQueue === true) return; // Eits if the cuntion is busy
+                node.lockHandleTelegramQueue = true; // Lock the function. It cannot be called again until finished.
+
+                // Retrieving oKNXMessage  { grpaddr, payload,dpt,outputtype (write or response),nodecallerid (node caller)}. 06/03/2020 "Read" request does have the lower priority in the queue, so firstly, i search for "read" telegrams and i move it on the top of the queue pile.
+                var aTelegramsFiltered = [];
+                aTelegramsFiltered = node.telegramsQueue.filter(a => a.outputtype !== "read");
+                if (aTelegramsFiltered.length == 0) {
+                    aTelegramsFiltered = node.telegramsQueue;
+                }
+                if (aTelegramsFiltered.length == 0) {
+                    node.lockHandleTelegramQueue = false; // Unlock the function
                     return;
                 }
-                // Retrieving oKNXMessage  { grpaddr, payload,dpt,outputtype (write or response)}
-                var oKNXMessage = node.telegramsQueue[node.telegramsQueue.length - 1]; // Get the last message in the queue
-                node.telegramsQueue.pop();// Remove the last message from the queue.
+
+                var oKNXMessage = aTelegramsFiltered[aTelegramsFiltered.length - 1]; // Get the last message in the queue
                 if (oKNXMessage.outputtype === "response") {
-                    node.knxConnection.respond(oKNXMessage.grpaddr, oKNXMessage.payload, oKNXMessage.dpt);
+                    try {
+                        node.knxConnection.respond(oKNXMessage.grpaddr, oKNXMessage.payload, oKNXMessage.dpt);
+                    } catch (error) {
+                        try {
+                            node.nodeClients.find(a => a.id === oKNXMessage.nodecallerid).setNodeStatus({ fill: "red", shape: "dot", text: "Send response " + error, payload: oKNXMessage.payload, GA: oKNXMessage.grpaddr, dpt: oKNXMessage.dpt, devicename: "" })
+                        } catch (error) { }
+                    }
                 } else if (oKNXMessage.outputtype === "read") {
-                    node.knxConnection.read(oKNXMessage.grpaddr);
+                    try {
+                        node.knxConnection.read(oKNXMessage.grpaddr);
+                    } catch (error) { }
                 } else {
-                    node.knxConnection.write(oKNXMessage.grpaddr, oKNXMessage.payload, oKNXMessage.dpt);
+                    try {
+                        node.knxConnection.write(oKNXMessage.grpaddr, oKNXMessage.payload, oKNXMessage.dpt);
+                    } catch (error) {
+                        try {
+                            node.nodeClients.find(a => a.id === oKNXMessage.nodecallerid).setNodeStatus({ fill: "red", shape: "dot", text: "Send write " + error, payload: oKNXMessage.payload, GA: oKNXMessage.grpaddr, dpt: oKNXMessage.dpt, devicename: "" })
+                        } catch (error) { }
+                    }
                 }
+                // Remove current item in the main node.telegramsQueue array
+                try {
+                    node.telegramsQueue = node.telegramsQueue.filter(item => {
+                        if (item !== oKNXMessage) {
+                            return item;
+                        }
+                    });
+                } catch (error) { }
+                node.lockHandleTelegramQueue = false; // Unlock the function
             }
         }
 
@@ -698,6 +731,10 @@ module.exports = (RED) => {
                                 if (node.stopETSImportIfNoDatapoint === "stop") {
                                     RED.log.error("knxUltimate: ABORT IMPORT OF ETS CSV FILE. To skip the invalid datapoint and continue import, change the related setting, located in the config node in the ETS import section.");
                                     return;
+                                } else {
+                                    // 02/03/2020 Whould you like to continue without datapoint? Good. Here a totally fake datapoint
+                                    RED.log.warn("knxUltimate: WARNING IMPORT OF ETS CSV FILE. Datapoint not set. You choosed to continue import. The Group Address will be imported with a fake datapoint 1.001. -> " + element.split("\t")[0] + " " + element.split("\t")[1]);
+                                    ajsonOutput.push({ ga: element.split("\t")[1], dpt: "1.001", devicename: sFather + element.split("\t")[0] + " (DPT NOT SET IN ETS - FAKE DPT USED)" });
                                 }
                             } else {
                                 var DPTa = element.split("\t")[5].split("-")[1];