+++ CHANGELOG.md
@@ -5,6 +5,23 @@
 <br/>
 
 <p>
+<b>Version 1.1.40</b> - March 2020<br/>
+- Better handling of telegrams, giving priority to the "write" and "response" telegram in the queue. Thanks @heleon19 for the suggestion.<br/>
+</p>
+<p>
+<b>Version 1.1.39</b> - March 2020<br/>
+- Fix a very low priority issue: a possible crash if you set the knx-ultimate node's output as "respond", while passing an object as payload to the input.<br/>
+</p>
+<p>
+<b>Version 1.1.38</b> - March 2020<br/>
+- Yet, if you import an ETS CSV file without datapoints, a fake datapoint 1.001 will be used (if you selected to force import the group address)<br/>
+- Update help and wiki to reflect the change<br/>
+</p>
+<p>
+<b>Version 1.1.37</b> - Feb 2020<br/>
+- Fixed an issue where new knx-ultimate nodes, without a gateway config node, doesn't open the configuration window. Thanks @svenflender<br/>
+</p>
+<p>
 <b>Version 1.1.36</b> - Feb 2020<br/>
 - You can now import ESF group address format, beside CSV.<br/>
 - Updated the Wiki.<br/>